- remove useless file utils
