+++ apps/data-scanner/lib/scripts/index.mjs
@@ -1,0 +1,1 @@
+export * from './updateSessions.mjs';
