+++ apps/data-scanner/lib/utils/index.mjs
@@ -1,2 +1,1 @@
-export * from './fileUtils.mjs';
 export * from './formatUtils.mjs';
